+++ CHANGELOG.md
@@ -3,3 +3,4 @@
   - target folder is now located at HOME
   - better serializators
   - error when java class does not exist
+  - added executor install command